Fixed Voltage Processing

+++ RED/Models/SensorModel.cs
@@ -8,7 +8,7 @@ namespace RED.Models
 {
     internal class SensorModel
     {
-        internal short Voltage;
+        internal float Voltage;
         internal byte Ultrasonic0;
         internal byte Ultrasonic1;
         internal byte Ultrasonic2;
